[Bronze II] Title: Hashing, Time: 72 ms, Memory: 12140 KB, Score: 50 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.kt
@@ -14,7 +14,7 @@ fun main() {
         repeat(i) {
             r *= 31
         }
-        total += (str[i] - 'a' + 1) * r
+        total += (str[i] - 'a' + 1).toLong() * r.toLong()
     }
 
     println(total % 1234567891)
